Use proper integer types in socket-related syscalls and ocalls
This contains two socket-related changes, which would be hard to factor
out into two commits: (because of compilation warnings)

	- Making ocalls use size_t in their interface. It's not the same as
	  Linux syscall interface anyway, so why not make it sane?
	- Fixing sycalls arguments types to match the ones use by kernel.
	  Note: there's no such type as "socklen_t" in the kernel, so it got
	  removed.

+++ Pal/src/host/Linux-SGX/db_pipes.c
@@ -122,7 +122,7 @@ static int pipe_listen(PAL_HANDLE* handle, const char* name, int options) {
         return -PAL_ERROR_DENIED;
 
     struct sockopt sock_options;
-    unsigned int addrlen = sizeof(struct sockaddr_un);
+    size_t addrlen = sizeof(struct sockaddr_un);
     int nonblock = options & PAL_OPTION_NONBLOCK ? SOCK_NONBLOCK : 0;
 
     ret = ocall_listen(AF_UNIX, SOCK_STREAM | nonblock, 0, /*ipv6_v6only=*/0,
